% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexsvg.R
\name{latexsvg}
\alias{latexsvg}
\title{Takes the code of an svg file that has <text> blocks with latex math
E.g. <text>\\(x^2\\)</text>
Renders the latex via MathJax and inlines it into the svg.
The function briefly starts a shinyapp since the transformation needs
MathJax to be run in the browser. Will work in the RStudio viewer
and in Firefox. Does not seem to work in Chrome.}
\usage{
latexsvg(svg = paste0(readLines(file), collapse = "\\n"), file = NULL,
  outfile = if (!is.null(file)) paste0(tools::file_path_sans_ext(file),
  "_latex.svg"))
}
\arguments{
\item{svg}{the code of svg as text}

\item{file}{a file name of the svg if no code is given}

\item{outfile}{if not null the created svg will be written to this file}
}
\value{
The code of the generated svg file
}
\description{
The javascript code that does the actually work is a minimally adapted
version of the code by Jason Sachs. See the following references:
}
\details{
https://www.embeddedrelated.com/showarticle/599.php
https://bitbucket.org/jason_s/svg_mathjax
}

