% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latexsvg.R
\name{pdf2svg}
\alias{pdf2svg}
\title{A function that converts a graphic in PDF format to an svg file
and saves the svg file in the same directory as the pdf file
Needs inkscape and the path to it binary}
\usage{
pdf2svg(file, inkscape.bin, svg.file = paste0(tools::file_path_sans_ext(file),
  ".svg"))
}
\arguments{
\item{file}{the pdf file}

\item{inkscape.bin}{path and file name of the inkscape binary}

\item{svg.file}{the output svg file; by default the same name as the pdf file but with extension .svg.}
}
\description{
A function that converts a graphic in PDF format to an svg file
and saves the svg file in the same directory as the pdf file
Needs inkscape and the path to it binary
}

